refactor: :recycle: move styles to styles folder, extract router to a separate file
