+++ index.html
@@ -2,9 +2,9 @@
 <html lang="en">
 	<head>
 		<meta charset="UTF-8" />
-		<link rel="icon" type="image/svg+xml" href="/vite.svg" />
+		<link rel="icon" type="image/svg+xml" href="/iphone.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Vite + React</title>
+		<title>Smarthub</title>
 	</head>
 	<body>
 		<div id="portal"></div>
