--- src/variables.scss
@@ -1,15 +1,0 @@
-:root {
-	--bg-primary: hsl(0, 0%, 100%);
-	--bg-secondary: hsla(0, 0%, 45%, 0.1);
-	--bg-tertiary: hsl(209, 65%, 47%);
-
-	--text-primary: hsl(0, 0%, 0%);
-	--text-secondary: hsl(0, 0%, 100%);
-	--text-tertiary: hsl(0, 0%, 45%);
-	--text-quaternary: hsl(202, 87%, 54%);
-
-	--h1-size: 2rem;
-
-	--btn-bg: hsl(202, 87%, 54%);
-	--select: hsla(0, 0%, 45%, 0.116);
-}
